GUI/requirements.txt
numpy==1.26.4
PyOpenGL==3.1.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
